django_ecommerce/vip_customer_service/requirements.txt
Django>=3.2.0,<4.0.0
djangorestframework>=3.12.0,<4.0.0
mysqlclient>=2.0.0
requests>=2.25.0 